chardev: Let IOReadHandler use size_t type
The number of bytes can not be negative nor zero.

?Qemu has integer overflows because IOReadHandler and its associated functions use a signed integer data type for a size value.

Fixed 2 format string:
- hw/char/spapr_vty.c
- hw/usb/ccid-card-passthru.c

Fixes: CVE-2018-18438
Reported-by: Arash Tohidi Chafi
Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Acked-by: Alberto Garcia <berto@igalia.com>

+++ hw/char/nrf51_uart.c
@@ -214,7 +214,7 @@ static void nrf51_uart_reset(DeviceState *dev)
     s->enabled = false;
 }
 
-static void uart_receive(void *opaque, const uint8_t *buf, int size)
+static void uart_receive(void *opaque, const uint8_t *buf, size_t size)
 {
 
     NRF51UARTState *s = NRF51_UART(opaque);
